src/requirements.txt
aniso8601==9.0.1
autopep8==2.0.2
bcrypt==4.0.1
bleach==6.0.0
blinker==1.5
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-mailman==0.3.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mkdocs-material-extensions==1.1.1
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.3.0
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.6
SQLAlchemy-Utils==0.40.0
tomli==2.0.1
typing_extensions==4.5.0
webencodings==0.5.1
Werkzeug==2.2.3
WTForms==3.0.1
